Move the concealEmail controller into its own sub-module
