+++ src/controllers/concealEmail/concealEmail_test.go
@@ -1,4 +1,4 @@
-package controllers
+package concealEmail
 
 import (
 	"github.com/halprin/email-conceal/src/context"
