+++ src/external/restApplicationContext.go
@@ -2,7 +2,7 @@ package external
 
 import (
 	"github.com/halprin/email-conceal/src/context"
-	"github.com/halprin/email-conceal/src/controllers"
+	"github.com/halprin/email-conceal/src/controllers/concealEmail"
 	"github.com/halprin/email-conceal/src/external/lib"
 	"github.com/halprin/email-conceal/src/gateways"
 	"github.com/halprin/email-conceal/src/usecases"
@@ -12,8 +12,8 @@ func init() {
 	var applicationContext = context.ApplicationContext{}
 
 	//controllers
-	applicationContext.Bind(func() controllers.ConcealEmailController {
-		return controllers.ConcealEmailController{}
+	applicationContext.Bind(func() concealEmail.ConcealEmailController {
+		return concealEmail.ConcealEmailController{}
 	})
 
 	//usecases
